Data Scraping/reddit_data_api/requirements.txt
requests
pandas
matplotlib
openpyxl
python-dotenv
praw
textblob